+++ collections/values_test.go
@@ -3,6 +3,10 @@ package collections
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/NibiruChain/nibiru/x/testutil"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	"github.com/gogo/protobuf/types"
@@ -16,5 +20,23 @@ func TestProtoValueEncoder(t *testing.T) {
 		cdc := codec.NewProtoCodec(registry)
 
 		assertValueBijective[types.BytesValue](t, ProtoValueEncoder[types.BytesValue](cdc), protoType)
+	})
+}
+
+func TestDecValueEncoder(t *testing.T) {
+	t.Run("bijectivity", func(t *testing.T) {
+		assertValueBijective(t, DecValueEncoder, sdk.MustNewDecFromStr("-1000.5858"))
+	})
+}
+
+func TestAccAddressValueEncoder(t *testing.T) {
+	t.Run("bijectivity", func(t *testing.T) {
+		assertValueBijective(t, AccAddressValueEncoder, testutil.AccAddress())
+	})
+}
+
+func TestUint64ValueEncoder(t *testing.T) {
+	t.Run("bijectivity", func(t *testing.T) {
+		assertValueBijective(t, Uint64ValueEncoder, 1000)
 	})
 }