+++ x/oracle/keeper/vote_target_test.go
@@ -3,17 +3,21 @@ package keeper
 import (
 	"testing"
 
+	"github.com/NibiruChain/nibiru/collections"
+
 	"github.com/stretchr/testify/require"
 )
 
 func TestKeeper_GetVoteTargets(t *testing.T) {
 	input := CreateTestInput(t)
 
-	input.OracleKeeper.ClearPairs(input.Ctx)
+	for _, p := range input.OracleKeeper.Pairs.Iterate(input.Ctx, collections.Range[string]{}).Keys() {
+		input.OracleKeeper.Pairs.Delete(input.Ctx, p)
+	}
 
 	expectedTargets := []string{"bar", "foo", "whoowhoo"}
 	for _, target := range expectedTargets {
-		input.OracleKeeper.SetPair(input.Ctx, target)
+		input.OracleKeeper.Pairs.Insert(input.Ctx, target)
 	}
 
 	targets := input.OracleKeeper.GetVoteTargets(input.Ctx)
@@ -23,11 +27,13 @@ func TestKeeper_GetVoteTargets(t *testing.T) {
 func TestKeeper_IsVoteTarget(t *testing.T) {
 	input := CreateTestInput(t)
 
-	input.OracleKeeper.ClearPairs(input.Ctx)
+	for _, p := range input.OracleKeeper.Pairs.Iterate(input.Ctx, collections.Range[string]{}).Keys() {
+		input.OracleKeeper.Pairs.Delete(input.Ctx, p)
+	}
 
 	validTargets := []string{"bar", "foo", "whoowhoo"}
 	for _, target := range validTargets {
-		input.OracleKeeper.SetPair(input.Ctx, target)
+		input.OracleKeeper.Pairs.Insert(input.Ctx, target)
 		require.True(t, input.OracleKeeper.IsVoteTarget(input.Ctx, target))
 	}
 }