+++ x/oracle/types/keys.go
@@ -1,10 +1,5 @@
 package types
 
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/address"
-)
-
 const (
 	// ModuleName is the name of the oracle module
 	ModuleName = "oracle"
@@ -18,82 +13,3 @@ const (
 	// QuerierRoute is the query router key for the oracle module
 	QuerierRoute = ModuleName
 )
-
-// Keys for oracle store
-// Items are stored with the following key: values
-//
-// - 0x01<pair_Bytes>: sdk.Dec
-//
-// - 0x02<valAddress_Bytes>: accAddress
-//
-// - 0x03<valAddress_Bytes>: int64
-//
-// - 0x04<valAddress_Bytes>: AggregateExchangeRatePrevote
-//
-// - 0x05<valAddress_Bytes>: AggregateExchangeRateVote
-//
-// - 0x06<pair_Bytes>: sdk.Dec
-var (
-	// Keys for store prefixes
-	ExchangeRateKey                 = []byte{0x01} // prefix for each key to a rate
-	FeederDelegationKey             = []byte{0x02} // prefix for each key to a feeder delegation
-	MissCounterKey                  = []byte{0x03} // prefix for each key to a miss counter
-	AggregateExchangeRatePrevoteKey = []byte{0x04} // prefix for each key to an aggregate prevote
-	AggregateExchangeRateVoteKey    = []byte{0x05} // prefix for each key to an aggregate vote
-	PairsKey                        = []byte{0x06} // prefix for each key to a pair
-	PairRewardsKey                  = []byte{0x07} // prefix for each key to a pair's rewards
-	PairRewardsCounterKey           = []byte{0x08} // prefix for the singleton of pair rewards ID.
-)
-
-// GetExchangeRateKey - stored by *pair*
-func GetExchangeRateKey(pair string) []byte {
-	return append(ExchangeRateKey, append([]byte(pair), 0x00)...)
-}
-
-// GetFeederDelegationKey - stored by *Validator* address
-func GetFeederDelegationKey(v sdk.ValAddress) []byte {
-	return append(FeederDelegationKey, address.MustLengthPrefix(v)...)
-}
-
-// GetMissCounterKey - stored by *Validator* address
-func GetMissCounterKey(v sdk.ValAddress) []byte {
-	return append(MissCounterKey, address.MustLengthPrefix(v)...)
-}
-
-// GetAggregateExchangeRatePrevoteKey - stored by *Validator* address
-func GetAggregateExchangeRatePrevoteKey(v sdk.ValAddress) []byte {
-	return append(AggregateExchangeRatePrevoteKey, address.MustLengthPrefix(v)...)
-}
-
-// GetAggregateExchangeRateVoteKey - stored by *Validator* address
-func GetAggregateExchangeRateVoteKey(v sdk.ValAddress) []byte {
-	return append(AggregateExchangeRateVoteKey, address.MustLengthPrefix(v)...)
-}
-
-// GetPairKey - stored by *pair* bytes
-func GetPairKey(d string) []byte {
-	return append(PairsKey, append([]byte(d), 0x00)...)
-}
-
-// ExtractPairFromPairKey - split pair from the pair key
-func ExtractPairFromPairKey(key []byte) (pair string) {
-	pair = string(key[1 : len(key)-1])
-	return
-}
-
-// GetPairRewardsKey returns the primary key for the PairRewards object.
-func GetPairRewardsKey(pair string, id uint64) []byte {
-	// TODO(mercilex): precompute key size
-	key := append(PairRewardsKey, []byte(pair)...)
-	key = append(key, 0x00)
-	key = append(key, sdk.Uint64ToBigEndian(id)...)
-	return key
-}
-
-// GetPairRewardsPrefixKey returns the primary key prefix
-// to iterate over the PairRewards instances of the pair.
-func GetPairRewardsPrefixKey(pair string) []byte {
-	key := append(PairRewardsKey, []byte(pair)...)
-	key = append(key, 0x00)
-	return key
-}